% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{clean_levels}
\alias{clean_levels}
\title{Clean org hierarchy levels}
\usage{
clean_levels(.df_levels)
}
\arguments{
\item{.df_levels}{Org levels as data frame}
}
\value{
df
}
\description{
Clean org hierarchy levels
}
\examples{
\dontrun{
  library(gisr)

  # Get PEPFAR Org Levels
  clean_levels(df_lvls)
 }

}
\keyword{internal}
