% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{clean_orgunits}
\alias{clean_orgunits}
\title{Clean Orgunits SQLView}
\usage{
clean_orgunits(.orgs, .org_levels)
}
\arguments{
\item{.orgs}{Raw country orgunits data from `datim_orgunits()`}

\item{.org_levels}{Expanded and Reshaped country orgunits levels}
}
\value{
Cleaned OU/Country Orgunits as a data frame
}
\description{
Clean Orgunits SQLView
}
\keyword{internal}
