% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{data_store}
\alias{data_store}
\title{Get and Set the Location of the Data Store}
\usage{
data_store(path = NULL)
}
\arguments{
\item{path}{an optional character string giving a new file path in which to
store the data}
}
\value{
a character string giving the current location of the data store if
  \code{path} was specified, this will be returned invisibly
}
\examples{
# list the current data store location
data_store()
old_store <- data_store()

# set it to a new value
data_store("~/a_new_path")
data_store()
}
