% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_dim_item}
\alias{datim_dim_item}
\title{Get dimension / item id}
\usage{
datim_dim_item(
  dimension,
  name,
  username,
  password,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{dimension}{Dimension name}

\item{name}{Item name}

\item{username}{DATIM Account Username}

\item{password}{DATIM Account password}

\item{baseurl}{DATIM API end point}
}
\value{
UID of item
}
\description{
Get dimension / item id
}
\examples{
\dontrun{
  library(grabr)

  datim_dim_item(dimension = "Funding Agency", name = "USAID")

  datim_dim_item(dimension = "Targets / Results", name = "MER Results")
  datim_dim_item(dimension = "Targets / Results", name = "MER Targets")
}

}
