% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_dim_items}
\alias{datim_dim_items}
\title{Get PEPFAR/DATIM Dimension Items}
\usage{
datim_dim_items(
  dimension,
  username,
  password,
  var = NULL,
  fields = NULL,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{dimension}{Dimension name}

\item{username}{DATIM Account Username}

\item{password}{DATIM Account passward}

\item{var}{column name to pull values from, id or item}

\item{fields}{list of column names to return, this will overwrite `var`}

\item{baseurl}{DATIM API end point}
}
\value{
Dimension's items as tibble or vector
}
\description{
Get PEPFAR/DATIM Dimension Items
}
\examples{
\dontrun{
  library(grabr)

  datim_dim_items(dimension = "Funding Agency")
  datim_dim_items(dimension = "Funding Agency", var = "item")
}

}
