% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_dimension}
\alias{datim_dimension}
\title{Get PEPFAR/DATIM Dimension ID}
\usage{
datim_dimension(name, username, password, baseurl = "https://final.datim.org/")
}
\arguments{
\item{name}{Dimension name}

\item{username}{DATIM Account Username, recommended using glamr::datim_user()`}

\item{password}{DATIM Account passward, recommended using glamr::datim_pwd()`}

\item{baseurl}{DATIM API End point, default value is `https://final.datim.org/`}
}
\value{
dimension uid
}
\description{
Get PEPFAR/DATIM Dimension ID
}
\examples{
\dontrun{
  library(grabr)
  datim_dimension("OU Level")
}

}
