% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_dimensions}
\alias{datim_dimensions}
\title{Get PEPFAR/DATIM dimensions}
\usage{
datim_dimensions(
  username,
  password,
  var = NULL,
  base_url = "https://final.datim.org/"
)
}
\arguments{
\item{username}{DATIM Account Username}

\item{password}{DATIM Account passward}

\item{var}{Column name to pull all values from, default is NULL, options are: id, dimension}

\item{base_url}{DATIM API End point, , default value is `https://final.datim.org/`}
}
\value{
Dimensions as tibble or list of ids / dimension names
}
\description{
Get PEPFAR/DATIM dimensions
}
\examples{
\dontrun{
  library(grabr)

  datim_dimensions()
}

}
