% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_orgunits}
\alias{datim_orgunits}
\title{Pull Orgunits SQLView}
\usage{
datim_orgunits(cntry, username, password, base_url = NULL)
}
\arguments{
\item{cntry}{Country name}

\item{username}{Datim username}

\item{password}{Datim password}

\item{base_url}{Datim API Base URL}
}
\value{
OU Orgunit as data frame
}
\description{
Pull Orgunits SQLView
}
\examples{
\dontrun{
  library(grabr)

  datim_orgunits(
   cntry = "Mozambique",
   username = glamr::datim_user(),
   password = glamr::datim_pwd()
  )
}
}
