% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_hierarchy.R
\name{datim_pull_hierarchy}
\alias{datim_pull_hierarchy}
\title{Extract PEPFAR Org Hierarchy}
\usage{
datim_pull_hierarchy(
  ou_uid,
  username,
  password,
  baseurl = "https://final.datim.org/",
  folderpath_output = NULL
)
}
\arguments{
\item{ou_uid}{UID for the country, recommend using `identify_ouuids()`}

\item{username}{DATIM username}

\item{password}{DATIM password, recommend using `mypwd()`}

\item{baseurl}{API base url, default = https://final.datim.org/}

\item{folderpath_output}{provide the full path to the folder for saving}
}
\description{
Extract PEPFAR Org Hierarchy
}
\note{
This function is migrated from `Wavelength::pull_hierarchy()` and is similar to `datim_orgunits()``
}
\examples{
\dontrun{
#get OU UID
  ouuid <- identify_ouuids() \%>\% dplyr::filter(ou == "Kenya")
#pull hierarchy (paths are all UIDs)
  df <- datim_pull_hierarchy(ouuid, username = myuser, password = mypwd(myuser)) }
}
