% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_sqlviews}
\alias{datim_sqlviews}
\title{Query Datim SQLViews}
\usage{
datim_sqlviews(
  username,
  password,
  view_name = NULL,
  dataset = FALSE,
  datauid = NULL,
  query = NULL,
  baseurl = NULL
)
}
\arguments{
\item{username}{Datim username}

\item{password}{Datim password}

\item{view_name}{Datim SQLView name}

\item{dataset}{Return SQLView dataset or uid? Default is false}

\item{datauid}{Data UID}

\item{query}{SQLView Query params, a list containing type and params key value pairs}

\item{baseurl}{Datim API Base URL}
}
\value{
SQLView uid or dataset as data frame
}
\description{
Query Datim SQLViews
}
\note{
This function should be used to identify Datim SQLView and Extract Data
}
\examples{
\dontrun{
  library(grabr)

  datim_sqlviews(
    username =glamr::datim_user(),
    password =glamr::datim_pwd(),
    view_name = "A list of OUs",
    dataset = TRUE
  )
}

}
