% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{.my_package_env}
\alias{.my_package_env}
\title{grabr environment for storing token}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.my_package_env
}
\description{
grabr environment for storing token
}
\keyword{datasets}
