% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_wave.R
\name{extract_cntry}
\alias{extract_cntry}
\title{Extract Country Name}
\usage{
extract_cntry(request_body)
}
\arguments{
\item{request_body}{elements to pass into the PDAP Wave POST API}
}
\value{
vector country's name
}
\description{
Extract Country Name
}
\keyword{internal}
