% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{extract_datim}
\alias{extract_datim}
\title{DATIM Analytics API}
\usage{
extract_datim(url, username, password)
}
\arguments{
\item{url}{supply url for API call}

\item{username}{DATIM Username, defaults to using glamr::datim_user()` if blank}

\item{password}{DATIM password, defaults to using glamr::datim_pwd()` if blank}
}
\value{
API pull of data from DATIM
}
\description{
DATIM Analytics API
}
\note{
Consider using glamr::datim_process_query()` and/or glamr::datim_execute_query()`
}
\examples{
\dontrun{
 myurl <- paste0(baseurl, "api/29/analytics.json?
                 dimension=LxhLO68FcXm:udCop657yzi&
                 dimension=ou:LEVEL-4;HfVjCurKxh2&
                 filter=pe:2018Oct&
                 displayProperty=SHORTNAME&outputIdScheme=CODE")

 df_targets <- extract_datim(myurl,glamr::datim_user(),glamr::datim_pwd())

 }

}
\keyword{internal}
