% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_wave.R
\name{flattenbody}
\alias{flattenbody}
\title{Flatten Body}
\usage{
flattenbody(x)
}
\arguments{
\item{x}{list}
}
\description{
A query can only have one value per name, so take
any values that contain vectors length >1 and
split them up. For example list(x=1:2, y="a") becomes list(x=1, x=2, y="a").
This code was developed and shard by Derek Wood (GHSD/PRIME)
}
\keyword{internal}
