% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{get_levels}
\alias{get_levels}
\title{Get all orgunits levels in org hierarchy}
\usage{
get_levels(
  username,
  password,
  expand = FALSE,
  reshape = FALSE,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{username}{DATIM username, recommed using glamr::datim_user()`}

\item{password}{DATIM password, recommend using glamr::datim_pwd()`}

\item{expand}{Fill in missing snu1 level? default is FALSE}

\item{reshape}{Reshape data as long? default is FALSE}

\item{baseurl}{base API url, default = https://final.datim.org/}
}
\value{
df
}
\description{
Get all orgunits levels in org hierarchy
}
\note{
Similar to `grabr::identify_levels()` and `grabr::get_outable()`
}
\examples{
\dontrun{
  library(grabr)

  # Get PEPFAR Org Levels
  get_levels()
 }

}
