% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{get_ouuid}
\alias{get_ouuid}
\title{Get Operatingunit / Country Org UID}
\usage{
get_ouuid(
  operatingunit,
  username,
  password,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{operatingunit}{Operatingunit name}

\item{username}{Datim Account username, recommend using glamr::datim_user()`}

\item{password}{Datim Account Password, recommend using glamr::datim_pwd()`}
}
\value{
uid
}
\description{
Get Operatingunit / Country Org UID
}
\examples{
\dontrun{
  library(grabr)

  # get orgunit for specific OU/Country: kenya
  get_ouuid(operatingunit = "Kenya")
}

}
