% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabr-package.R
\docType{package}
\name{grabr-package}
\alias{grabr}
\alias{grabr-package}
\title{grabr: OHA/SI APIs Package}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a series of base functions useful to the GH OHA SI team. These function extend the utility functions in glamr, focusing primarily on API utility functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USAID-OHA-SI/grabr}
  \item \url{https://usaid-oha-si.github.io/grabr/}
  \item Report bugs at \url{https://github.com/USAID-OHA-SI/grabr/issues}
}

}
\author{
\strong{Maintainer}: Aaron Chafetz \email{achafetz@usaid.gov}

Authors:
\itemize{
  \item Tim Essam \email{tessam@usaid.gov}
  \item Baboyma Kagniniwa \email{bkagniniwa@usaid.gov}
}

}
\keyword{internal}
