% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_hierarchy.R
\name{hierarchy_extract}
\alias{hierarchy_extract}
\title{Pull Hierarchy Data from DATIM}
\usage{
hierarchy_extract(
  ou_uid,
  username,
  password,
  add_geom = FALSE,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{ou_uid}{UID for the country, recommend using identify_ouuids}

\item{username}{DATIM username, recommend using datim_user}

\item{password}{DATIM password, recommend using datim_pwd}

\item{add_geom}{Add geometry column to the output, default is false}

\item{baseurl}{API base url, default = https://final.datim.org/}
}
\description{
internal funciton to extract orgunit hierarchy
}
\note{
Migratated from Wavelength
}
\examples{
\dontrun{
#get OU UID
  ouuid <- identify_ouuids() \%>\% dplyr::filter(ou == "Kenya")
#pull hierarchy (paths are all UIDs)
  df <- hierarchy_extract(ouuid, username = myuser, password = datim_pwd(myuser))
  }
}
