% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_hierarchy.R
\name{hierarchy_identify_ctry}
\alias{hierarchy_identify_ctry}
\title{Extract country name from OU or country name}
\usage{
hierarchy_identify_ctry(df)
}
\arguments{
\item{df}{data frame}
}
\value{
Unique country names
}
\description{
Extract country name from OU or country name
}
\keyword{internal}
