% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_hierarchy.R
\name{hierarchy_rename}
\alias{hierarchy_rename}
\title{Rename Hierarchy}
\usage{
hierarchy_rename(
  df,
  country,
  username,
  password,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{df}{data frame created by hierarchy_extract and hierarchy_clean}

\item{country}{county name, eg "Malawi" or "Nepal"}

\item{username}{DATIM username, recommend using datim_user}

\item{password}{DATIM password, recommend using datim_pwd}

\item{baseurl}{API base url, default = https://final.datim.org/}
}
\value{
Cleaned/Renamed data
}
\description{
Internal function to Rename Hierarchy
}
\note{
Rename from Levels to OU/SNU1/PSNU/Facility
}
