% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lazy_secrets}
\alias{lazy_secrets}
\title{Lazy checking/loading of credentials}
\usage{
lazy_secrets(service = c("datim", "pano", "pdap", "s3"), username, password)
}
\arguments{
\item{service}{account, either datim", "pano", "pdap", or "s3"}

\item{username}{account username or s3 access key}

\item{password}{account password or s3 secret key}
}
\value{
returns a list of 2 - username/access and password/secret
}
\description{
This function is useful within another function. It check whether a username
or password has been provided by the user and, if not, checks if they were
stored via `glamr` or prompts user to provide credentials through interactive
prompt.
}
\examples{
\dontrun{
  accnt <- lazy_secrets("datim", username = username, password = password)
  datim_dimensions(accnt$username, accnt$password)
}
}
