% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{package_check}
\alias{package_check}
\title{Check if package exists}
\usage{
package_check(pkg)
}
\arguments{
\item{pkg}{package name}
}
\value{
warning message if package is not installed
}
\description{
Check if package exists
}
\seealso{
Other utility: 
\code{\link{get_baseurl}()},
\code{\link{var_exists}()}
}
\concept{utility}
