% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_extract}
\alias{pano_extract}
\title{Extract data outputs from Panorama}
\usage{
pano_extract(
  item = "mer",
  version,
  fiscal_year,
  quarter,
  unpack = FALSE,
  username,
  password,
  base_url = "https://pepfar-panorama.org"
)
}
\arguments{
\item{item}{Panorama data type. Eg: mer, financial, sims, narratives}

\item{version}{Data release version: Initial or Clean, defaults to
current version if blank}

\item{fiscal_year}{Reporting Fiscal year, defaults to current FY if blank}

\item{quarter}{Reporting Quarter, defaults to current quarter if blank}

\item{unpack}{If TRUE, unpack nested directories}

\item{username}{Panorama username, recommend using `glamr::pano_user()`}

\item{password}{Panorama password, recommend using `glamr::pano_pwd()`}

\item{base_url}{Panorama base url}
}
\value{
list of output files as data frame
}
\description{
Extract data outputs from Panorama
}
\note{
This function combines `pano_session()`, `pano_content()`, `pano_elements()`, and in some cases `pano_unpack()`
}
\examples{
\dontrun{
  library(tidyverse)
  library(grabr)

  pano_extract(item = "mer")
}

}
