% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_extract_msd}
\alias{pano_extract_msd}
\title{Downloads Country Specific MSDs}
\usage{
pano_extract_msd(
  operatingunit = NULL,
  version,
  fiscal_year,
  quarter,
  level = c("psnu", "ou", "site", "nat"),
  dest_path,
  username,
  password
)
}
\arguments{
\item{operatingunit}{PEPFAR Operating Unit. Default is set to NULL for
to return global datasets}

\item{version}{Data release version: "initial" or "clean", defaults to
current version}

\item{fiscal_year}{Reporting Fiscal year, defaults to current version}

\item{quarter}{Reporting Quarter: Single digit quarters, defaults to
current version}

\item{level}{Org level, options are psnu" (default), "ou", "site",
or "nat"}

\item{dest_path}{Directory path to download file. Default set to
`si_path()`}

\item{username}{Panorama username, recommend using `glamr::pano_user()`,
which is the default if left blank}

\item{password}{Panorama password, recommend using `glamr::pano_pwd()`,
which is the default if left blank}
}
\description{
Downloads Country Specific MSDs
}
\examples{
\dontrun{
 pano_extract_msd(operatingunit = "Zambia", level = "site")
}

}
