% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{PEPFAR Panorama base url}
\alias{PEPFAR Panorama base url}
\alias{pano_url}
\title{PEPFAR Panorama Base URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
pano_url
}
\description{
PEPFAR Panorama Base URL
}
\keyword{internal}
