% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_mer.R
\name{pull_mer}
\alias{pull_mer}
\title{Extract DATIM Results and Targets (DATIM API Call)}
\usage{
pull_mer(
  ou_name = NULL,
  username,
  password,
  baseurl = "https://final.datim.org/",
  fy_pd = NULL,
  quarters_complete = NULL,
  folderpath_output = NULL
)
}
\arguments{
\item{ou_name}{Operating Unit name, if mechanism is not specified}

\item{username}{DATIM username}

\item{password}{DATIM password, recommend using `mypwd()`}

\item{baseurl}{API base url, default = https://final.datim.org/}

\item{fy_pd}{fiscal year(s) to cover, default will be current FY if not provided}

\item{quarters_complete}{no. of quarters completed through FY to determine weeks left in year}

\item{folderpath_output}{folder path to store DATIM output, default = NULL}
}
\description{
Extract DATIM Results and Targets (DATIM API Call)
}
\examples{
\dontrun{
 #ou mer data
 myuser <- "UserX"
 mech_x_dta <- pull_mer(ou_name = "Namibia", username = myuser, password = mypwd(myuser))
 }
}
