% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_mer.R
\name{update_meta_mer}
\alias{update_meta_mer}
\title{Update MER meta table}
\usage{
update_meta_mer(fy_pd = NULL, savefolder = "out/DATIM", upload = FALSE)
}
\arguments{
\item{fy_pd}{fiscal year(s) to cover, default will be current FY if not provided}

\item{savefolder}{folderpath to save, default = "out/DATIM"}

\item{upload}{should the new table be pushed to s3? default = FALSE}
}
\description{
Update MER meta table
}
