% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_hierarchy.R
\name{update_meta_orgs}
\alias{update_meta_orgs}
\title{Update Organization Hierarchy meta table}
\usage{
update_meta_orgs(savefolder = "out/DATIM", upload = FALSE)
}
\arguments{
\item{savefolder}{folderpath to save, default = "out/DATIM"}

\item{upload}{should the new table be pushed to Google Drive and s3? default = FALSE}
}
\description{
Update Organization Hierarchy meta table
}
