% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_mer.R
\name{update_meta_targets}
\alias{update_meta_targets}
\title{Update MER targets from MSD (prior to site results being available)}
\usage{
update_meta_targets(fy, savefolder = "out/DATIM", upload = FALSE)
}
\arguments{
\item{fy}{fiscal year}

\item{savefolder}{folderpath to save, default = "out/DATIM"}

\item{upload}{should the new table be pushed to s3? default = FALSE}
}
\description{
Update MER targets from MSD (prior to site results being available)
}
