% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_wave.R
\name{wave_check_token}
\alias{wave_check_token}
\title{Check if PDAP Wave token is already established}
\usage{
wave_check_token(timeout = 20)
}
\arguments{
\item{timeout}{time session is valid for, default = 20}
}
\value{
token (if it exists and is valid)
}
\description{
Check if PDAP Wave token is already established
}
\keyword{internal}
