% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_wave.R
\name{wave_store_token}
\alias{wave_store_token}
\title{Store token and time to package environment}
\usage{
wave_store_token(token)
}
\arguments{
\item{token}{PDAP Wave token}
}
\value{
store token and time to to package environment
}
\description{
Store token and time to package environment
}
\keyword{internal}
