% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{clean_orgunits}
\alias{clean_orgunits}
\title{Clean Orgunits SQLView}
\usage{
clean_orgunits(.orgs, cntry, username, password, baseurl)
}
\arguments{
\item{.orgs}{Raw data from `datim_orgunits()`}

\item{cntry}{Country name}

\item{username}{Datim username}

\item{password}{Datim password}

\item{baseurl}{Datim API Base URL}
}
\value{
OU/Country Orgunits as a data frame
}
\description{
Clean Orgunits SQLView
}
