% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_orgunits}
\alias{datim_orgunits}
\title{Pull Orgunits SQLView}
\usage{
datim_orgunits(
  cntry,
  username,
  password,
  reshape = FALSE,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{cntry}{Country name}

\item{username}{Datim username}

\item{password}{Datim password}

\item{reshape}{Unpack parent org units}

\item{baseurl}{Datim API Base URL, default to https://final.datim.org/}
}
\value{
OU/Country Orgunits as a data frame
}
\description{
Pull Orgunits SQLView
}
\examples{
\dontrun{
  library(grabr)

  datim_orgunits(
   cntry = "Mozambique",
   username = glamr::datim_user(),
   password = glamr::datim_pwd()
  )
}
}
