% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_wave.R
\name{est_export_path}
\alias{est_export_path}
\title{Establish Export Path}
\usage{
est_export_path(request_body, folderpath_dwnld, psd_type)
}
\arguments{
\item{request_body}{elements to pass into the PDAP Wave POST API}

\item{folderpath_dwnld}{where to download, default = "Data"}

\item{psd_type}{Type of PEPFAR Structured dataset: "psnu_im" (default),
"ou_im", or "site_im"}
}
\value{
vector of the filename
}
\description{
Establish Export Path
}
\keyword{internal}
