% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_baseurl}
\alias{get_baseurl}
\title{Get base url from a link}
\usage{
get_baseurl(url)
}
\arguments{
\item{url}{DATIM API end points}
}
\value{
Base url without trailing slash
}
\description{
Get base url from a link
}
\seealso{
Other utility: 
\code{\link{var_exists}()}
}
\concept{utility}
