% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_download}
\alias{pano_download}
\title{Download file from PEPFAR Panorama}
\usage{
pano_download(
  item_url,
  username,
  password,
  session = NULL,
  dest_path = NULL,
  uncompress = FALSE
)
}
\arguments{
\item{item_url}{URL for the item to be downlaoded}

\item{username}{Username for PEPFAR Panorama Account. Recommend using `pano_user()`}

\item{password}{Password for PEPFAR Panorama Account. Recommend using `pano_pwd()`}

\item{session}{Login session, only used within other `pano_extract_*()`}

\item{dest_path}{Location and name of the destination file}

\item{uncompress}{If yes, the downloaded zip file will be decompressed. Default is FALSE}
}
\value{
file content as binary
}
\description{
Download file from PEPFAR Panorama
}
\examples{
\dontrun{
  library(tidyverse)
  library(grabr)
  library(glamr)

  url <- "https://pepfar-panorama.org/forms/downloads"

  elts <- pano_items(page_html = url,
                     username = pano_user(), password = pano_pwd())

  f_url <- elts \%>\% filter(type == "file zipfile") \%>\% pull(path) \%>\% first()

  pano_download(item_url = url, session = s, dest = "Data") }

}
