% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_wave.R
\name{wave_est_session}
\alias{wave_est_session}
\title{Establish PDAP Wave Session}
\usage{
wave_est_session(
  username,
  password,
  datim_hostname = "genie.testing.datim.org",
  api_host = "wave.test.pdap.pepfar.net"
)
}
\arguments{
\item{username}{DEPRECATED - DATIM username, if blank looks for stored credentials
(\code{glamr::set_datim()}) and then prompts for credentials if not found}

\item{password}{DEPRECATED -DATIM password, if blank looks for stored credentials
(\code{glamr::set_datim()}) and then prompts for credentials if not found}

\item{datim_hostname}{DATIM URL: 'genie.testing.datim.org'}

\item{api_host}{API address: 'wave.test.pdap.pepfar.net'}
}
\value{
session token
}
\description{
Establish PDAP Wave Session
}
\keyword{internal}
